Added game logic and required private methods to the MainActivity class.
Removed unnecessary code from the Card class

Added a comment to discardHand() in the PlayerHand class.
Added a boolean method to check if the player has a 5 card trick: isFiveCardTrick() to the PlayerHand class.

Added a raw recourse directory to hold audio files.
Added 9 audio files to be played when dealing cards to the raw directory.

Cleaned up whitespace and unused imports in the DeckTest class.

Added a test to the PlayerHandTest class (Testing double cards, such as two Seven of Hearts).